feat: support `conventional-changelog-*` presets
Support the ability to pass in `conventional-changelog-*` preset pakage
names using the `preset` option.

Previous behavior of the `preset` option was to map the name provided
through the option to one of the presets built into this
`conventional-recommended-bump` package.

Now we take the name passed to the `preset` option, whether passed programmatically
or via the CLI, append it to `conventional-changelog-`, and then attempt to
import that package.

The loading behavior matches the behavior of the `preset` option in
`conventional-changelog`.

Furthermore, the loading behavior has been extended to support the
different export types of conventional preset packages, such as a
promise, function, or object.

Lastly, we defer to `recommendedBumpOpts.parserOpts` property of a preset
package if it exists, as our initial refactor efforts created a
`parserOpts` object under `recommendedBumpOpts` in each preset package.

In the future we want to merge differences found in
`recommendedBumpOpts.parserOpts` into the top-level `parserOpts` object
and remove `recommendedBumpOpts.parserOpts` from each presetpackage if
it exists.

BREAKING CHANGE:

Will only accept the name associated with a `conventional-changelog-*`
preset package for the purpose of loading preset configuration.

Built-in presets previously available in `conventional-recommended-bump`
are no longer available.

Related to #241